fin du projet

+++ controllers/stuff.js
@@ -1,4 +1,6 @@
 const Thing = require("../models/Thing");
+const fs = require("fs");
+const { error } = require("console");
 
 // exports.createThing = (req, res, next) => {
 //     delete req.body._id;
@@ -53,10 +55,27 @@ exports.modifyThing = (req, res, next) => {
     .catch(error => res.status(400).json({ error }));
 };
 
+// exports.deleteThing = (req, res, next) => {
+//     Thing.deleteOne({ _id: req.params.id })
+//     .then(() => res.status(200).json({ message: "Objet suprimmé!" }))
+//     .catch(error => res.status(400).json({error}));
+// };
+
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet suprimmé!" }))
-    .catch(error => res.status(400).json({error}));
+    Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+        if (thing.userId != req.auth.userId) {
+            res.status(401).json({ message: "Non-authorisé" });
+        } else {
+            const filename = thing.imageUrl.split('/images/')[1];
+            fs.unlink(`image/${filename}`, () => {
+                Thing.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: "Objet suprimmé!" }))
+                .catch(error => res.status(400).json({ error }));
+            });
+        }
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.getOneThing = (req, res, next) => {
